requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
Flask_Cors==3.0.10
mysql_connector==0.3.1
python-dotenv==1.0.0
python_telegram_bot==13.14
requests==2.22.0
telegram==0.0.1
user-agents==2.2.0